requirements.txt
astroid==2.3.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.17.0
google-api-python-client==1.8.1
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
oauth2client==4.1.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.0
